Added User to the test process instead of greeting
Inserted User class into the android project.
This is a duplicate of the user class on server.
I then Changed Greeting for user in the android program. So a user was
sent and returned.

I changed the controller on the server to handle the User request -
taking two parameters - name and password. These were hardcoded into the
android url that was sent.

I changed the buttons string.xml to display the correct labels. Now
username and role were returned.

Ran the project and the user name was sent along with string password,
returning a User object - from this user i used onPostExecute to set the
text.