Added new buttons and send strings to url
Added text fields to type in, added submit button and added a new click
on method within MainActivity that activates the HttpRequestTask method
to send the rest request to the server. the method name is in the button
setup and relays directly to this method

It takes the string username and password and send it via post in the
url.

tested the connection - makes a connect accepts the password and then
shows the user.