Added second page and button to get there
Added a second page and work using the standard tutorial -
http://developer.android.com/training/basics/firstapp/starting-activity.html

once logged in then the user can click on choose file this will take
them to a new page where they can will be able to log out and select a
file to send to server.

This page is a child to the parent.
Made a new activity display result xml to layout the file.
Used the string .xml to label the new screen
added a new DisplayResultActivity.java class to control access and
objects in this new page