Added slightly different way to login
Added the use of HttpHeaders, where now they can take a
HttpAuthentication object - username and password and can be passed via
json to the server.

The server will identify this and use spring security to check username
and password. In turn the user can be logged in as any user on the web
browser version.

A Response entity is request  in return - sent is a HTTP>GET request.
on return of the entity, the user is removed from it by using get
body(); this is the user sent by the server.

ref:
https://github.com/spring-projects/spring-android-samples/blob/master/spring-android-basic-auth/client/src/org/springframework/android/basicauth/MainActivity.java
28/07/2015