Added bitmap resize and a upload message confirm
Used a method already provided on stack overflow - //REF:
http://stackoverflow.com/questions/6410364/how-to-scale-bitmap-to-screen-size
to resize the image to display on screen. this does it for all images
now, and no errors when uploading.

added correct messages to screen when file is or isnt uploaded. User can
work this out now if there is and issue.

added a new method to check at login the ip address and port are active.
If not then message is correctly displayed.
If username and or password are incorrect then the correct message is
displayed. rather than just saying denied.